feat(box): add box component
- box component
- add editorconfig, gitignore and prettierrc
