+++ .gitignore
@@ -1,4 +1,13 @@
 *.log
-/node_modules/
-/dist
-/lib
+node_modules/
+lerna-debug.log
+npm-debug.log
+packages/*/lib
+storybook-static
+packages/*/coverage
+packages/*/dist
+packages/*/node_modules
+npm-debug.log*
+packages/*/.env
+yarn.lock
+yarn-error.log
