--- packages/eDaN/src/index.js
@@ -1,5 +1,0 @@
-import React from 'react'
-
-export {
-    /* list all component */
-}
