--- packages/eDaN/src/theme/colors.js
@@ -1,21 +1,0 @@
-/*
- * @see {https://github.com/evilfactorylabsarchive/design#colors}
- *
- * 666 is key of light
- * 999 is key of dark
- */
-const colors = {
-    red: {
-        666: "#E53A40",
-        999: "#D01C22"
-    },
-    black: {
-        666: "#333333",
-        999: "#000000"
-    },
-    navy: "#001B44",
-    silver: "#eeeeee",
-    gray: "#dddddd",
-}
-
-export default colors
