--- packages/eDaN/src/theme/typography.js
@@ -1,9 +1,0 @@
-const fonts = {
-    heading: `Merriweather`,
-    body: `Inter`,
-    mono: `"Space Mono"`
-}
-
-export default {
-    fonts
-}
