+++ packages/edan-core/src/index.js
@@ -1,0 +1,1 @@
+export { default as Box } from './box'
