+++ packages/edan-core/src/theme/typography.js
@@ -1,0 +1,9 @@
+const fonts = {
+  heading: `Merriweather`,
+  body: `Inter`,
+  mono: `"Space Mono"`
+}
+
+export default {
+  fonts
+}
